% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{cigar}
\alias{cigar}
\title{Cigar}
\format{
\subsection{\code{cigar}}{

A data frame with 1,380 rows and 9 columns:
\describe{
\item{state}{State abbreviation}
\item{year}{Year}
\item{price}{Price per pack of cigarettes}
\item{population}{Population}
\item{population_16}{Population over the age of 16}
\item{cpi}{Consumer price index (1983 = 100)}
\item{ndi}{Per capita disposable income}
\item{sales}{Cigarette sales in packs per capita}
\item{pimin}{Minimum price in adjoining states per pack of cigarettes}
}
}
}
\source{
Baltagi and Levin (1992) and Baltagi, Griffin and Xiong (2000).
}
\usage{
cigar
}
\description{
Panel Data, 46 U.S. States over the period 1963-1992.
}
\keyword{datasets}
