% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{gasoline}
\alias{gasoline}
\title{Gasoline}
\format{
\subsection{\code{gasoline}}{

A data frame with 342 rows and 6 columns:
\describe{
\item{country}{Country}
\item{year}{Year}
\item{lgaspcar}{The logarithm of motor gasoline consumption per auto}
\item{lincomep}{The logarithm of real per capita income}
\item{lrpmg}{The logarithm of real motor gasoline price}
\item{lcarpcap}{The logarithm of the stock of cars per capita}
}
}
}
\source{
Baltagi and Griffin (1983).
}
\usage{
gasoline
}
\description{
Panel Data, 18 OECD countries over 19 years, 1960‑1978.
}
\keyword{datasets}
