% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{grunfeld}
\alias{grunfeld}
\title{Grunfeld}
\format{
\subsection{\code{grunfeld}}{

A data frame with 200 rows and 5 columns:
\describe{
\item{firm}{Firm number}
\item{year}{Year}
\item{investment}{Annual real gross investment}
\item{value}{Real value of the firm (shares outstanding)}
\item{capital}{Real value of the capital stock}
}
}
}
\source{
Grunfeld (1958).
}
\usage{
grunfeld
}
\description{
Panel Data, 10 U.S. firms over 20 years, 1935‑1954.
}
\keyword{datasets}
