% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{hedonic}
\alias{hedonic}
\title{Hedonic}
\format{
\subsection{\code{hedonic}}{

A data frame with 506 rows and 15 columns:
\describe{
\item{mv}{Medium value of owner-occupied homes}
\item{crim}{Crime rate}
\item{zn}{Proportion of 25,000 square feet residential lots}
\item{indus}{Proportion of non-retail business acres}
\item{chas}{Represents the dummy variable which is 1 if the tract bounds the Charles River}
\item{nox}{Annual average nitrogen oxide concentration in parts per hundred million}
\item{rm}{Average number of rooms}
\item{age}{Proportion of owner units built prior to 1940}
\item{dis}{Weighted distances to five employment centers in the Boston area}
\item{rad}{Index of accessibility to radial highways}
\item{tax}{Full value property tax rate ($/$10,000)}
\item{ptratio}{Pupil / teacher ratio}
\item{b}{Proportion of blacks in the population}
\item{lstat}{Proportion of the population that is lower status}
\item{townid}{Town identifier}
}
}
}
\source{
Harrison and Rubinfeld (1978). See also Belsley, Kuh and Welsch (1980).
}
\usage{
hedonic
}
\description{
Cross-Section, 506 Census tracts in the Boston area in 1970.
}
\keyword{datasets}
