% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{produc}
\alias{produc}
\title{Produc}
\format{
\subsection{\code{produc}}{

\describe{
\item{state}{State}
\item{st_abb}{State abbreviation}
\item{year}{Year}
\item{pub_cap}{Public capital}
\item{hwy}{Highway and streets}
\item{water}{Water and sewer facilities}
\item{util}{Other public buildings and structures}
\item{priv_cap}{Private capital stock}
\item{gsp}{Gross state products}
\item{emp}{Labor input measured by the employment in non-agricultural payrolls}
\item{unemp}{State unemployment rate}
}
}
}
\source{
Munnell (1990), see also Baltagi and Pinnoi (1995).
}
\usage{
produc
}
\description{
Panel Data, 48 states over 17 years, 1970-1986.
}
\keyword{datasets}
