% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltagi-package.R
\docType{data}
\name{wages}
\alias{wages}
\title{Wages}
\format{
\subsection{\code{wages}}{

\describe{
\item{exp}{Years of full‑time work experience}
\item{wks}{Weeks worked}
\item{occ}{OCC=1, if the individual is in a blue‑collar occupation}
\item{ind}{IND=1, if the individual works in a manufacturing industry}
\item{south}{SOUTH=1, if the individual resides in the South}
\item{smsa}{SMSA=1, if the individual resides in a standard metropolitan statistical area}
\item{ms}{MS=1, if the individual is married}
\item{fem}{FEM=1, if the individual is female}
\item{union}{UNION=1, if the individual's wage is set by a union contract}
\item{ed}{Years of education}
\item{blk}{BLK=1, if the individual is black}
\item{lwage}{Logarithm of wage}
}
}
}
\source{
The Panel Study of Income Dynamics, taken from Cornwell and Rupert (1988).
}
\usage{
wages
}
\description{
Panel Data, 595 individuals over 7 years, 1976-1982.
}
\keyword{datasets}
